+++ lib/widget/nothing_display.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+// import '../const.dart';
+
+class NothingDisplay extends StatelessWidget {
+  const NothingDisplay();
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: const Text('該当するユーザーがいません'),
+      ),
+    );
+  }
+}
